requirements.txt
torchsummary==1.5.1
tqdm
tensorboard
tensorboardx
pandas
efficientnet_pytorch
torch_optimizer